[refac] apply facade pattern
